requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113

numpy==1.23.4
python-igraph==0.10.2
tqdm==4.64.1
torch==1.12.1+cu113
scipy==1.9.3
